Replace Subcomponent with Component in data:repository module.
