+++ app/src/main/java/jp/kuaddo/tsuidezake/TsuidezakeApp.kt
@@ -3,6 +3,7 @@ package jp.kuaddo.tsuidezake
 import com.jakewharton.threetenabp.AndroidThreeTen
 import dagger.android.AndroidInjector
 import dagger.android.DaggerApplication
+import jp.kuaddo.tsuidezake.data.repository.DaggerRepositoryComponent
 import jp.kuaddo.tsuidezake.di.DaggerAppComponent
 import timber.log.Timber
 
@@ -16,8 +17,11 @@ class TsuidezakeApp : DaggerApplication() {
         }
     }
 
-    override fun applicationInjector(): AndroidInjector<out DaggerApplication> {
-        return DaggerAppComponent.factory()
-            .create(this, applicationContext)
-    }
+    override fun applicationInjector(): AndroidInjector<out DaggerApplication> =
+        DaggerAppComponent.factory()
+            .create(
+                this,
+                applicationContext,
+                DaggerRepositoryComponent.factory().create(applicationContext)
+            )
 }
