+++ app/src/main/java/jp/kuaddo/tsuidezake/di/AppComponent.kt
@@ -6,7 +6,7 @@ import dagger.Component
 import dagger.android.AndroidInjector
 import dagger.android.support.AndroidSupportInjectionModule
 import jp.kuaddo.tsuidezake.TsuidezakeApp
-import jp.kuaddo.tsuidezake.data.repository.RepositoryModule
+import jp.kuaddo.tsuidezake.data.repository.RepositoryComponent
 import jp.kuaddo.tsuidezake.di.module.ActivityModule
 import jp.kuaddo.tsuidezake.di.module.AppModule
 import jp.kuaddo.tsuidezake.di.module.ViewModelDelegateModule
@@ -20,17 +20,19 @@ import javax.inject.Singleton
         ActivityModule::class,
         ViewModelModule::class,
         ViewModelDelegateModule::class,
-        AppModule::class,
-        RepositoryModule::class
+        AppModule::class
+    ],
+    dependencies = [
+        RepositoryComponent::class
     ]
 )
 interface AppComponent : AndroidInjector<TsuidezakeApp> {
-
     @Component.Factory
     interface Factory {
         fun create(
             @BindsInstance application: TsuidezakeApp,
-            @BindsInstance applicationContext: Context
+            @BindsInstance applicationContext: Context,
+            repositoryComponent: RepositoryComponent
         ): AppComponent
     }
