+++ data/repository/src/main/java/jp/kuaddo/tsuidezake/data/repository/internal/di/RepositoryModule.kt
@@ -1,0 +1,12 @@
+package jp.kuaddo.tsuidezake.data.repository.internal.di
+
+import dagger.Binds
+import dagger.Module
+import jp.kuaddo.tsuidezake.data.repository.Repository
+import jp.kuaddo.tsuidezake.data.repository.internal.RepositoryImpl
+
+@Module
+internal abstract class RepositoryModule {
+    @Binds
+    abstract fun bindRepository(impl: RepositoryImpl): Repository
+}
